+++ NEWS.md
@@ -28,6 +28,7 @@
 ### Server
 
 * Include active project in document title (caption of browser tab) 
+* Quit session command now accessible from global toolbar
 * Server Pro: Support for multiple concurrent R sessions per-user
 * Server Pro: Don't close PAM sessions by default (configurable via an option)